Rename normaliseDictEntries.ts
